Added a gloabl meta data index
* so that one can restore files from multiple Backups at the same time
* made the reference index an uint, because when someone would backup hunders of locations per day, after only two years the old solution would fall flat

+++ Backup.Shared/BackupEncryptionHelper.cs
@@ -88,7 +88,7 @@ public class BackupEncryptionHelper
         }
     }
 
-    public static void ConvertToFileNodes((string, DateTime)[] dataLines, ushort index, List<FileNode> returnValues)
+    public static void ConvertToFileNodes((string, DateTime)[] dataLines, uint index, List<FileNode> returnValues)
     {
         for (int d = 0; d < dataLines.Length; d++)
         {
